judge-api/requirements.txt
django
djangorestframework
markdown
django-filter
pygments
coreapi
django-cors-headers
djangorestframework-camel-case
django-query-builder
psycopg2